Made read() use 64-bit position rather than 32 bit to support large files

+++ RNFSManager.m
@@ -323,7 +323,7 @@ RCT_EXPORT_METHOD(read:(NSString *)filepath
     }
 
     // Seek to the position if there is one.
-    [file seekToFileOffset: (int)position];
+    [file seekToFileOffset: (long)position];
 
     NSData *content;
     if ((int)length > 0) {
